+++ src/paco/models/loader.py
@@ -35,7 +35,8 @@ from paco.models.applications import Application, PinpointApplication, ResourceG
     RDSOptionConfiguration, RDSClusterInstance, RDSClusterInstances, RDSClusterDefaultInstance, \
     DeploymentPipeline, DeploymentPipelineConfiguration, DeploymentPipelineSourceStage, DeploymentPipelineBuildStage, \
     DeploymentPipelineDeployStage, DeploymentPipelineSourceCodeCommit, DeploymentPipelineBuildCodeBuild, \
-    ECRRepositoryPermission, \
+    ECRRepositoryPermission, DeploymentPipelineBuildReleasePhase, DeploymentPipelineBuildReleasePhases, \
+    DeploymentPipelineBuildReleasePhaseCommand, \
     DeploymentPipelineDeployCodeDeploy, DeploymentPipelineManualApproval, CodeDeployMinimumHealthyHosts, \
     DeploymentPipelineDeployS3, DeploymentPipelineLambdaInvoke, DeploymentPipelineSourceGitHub, DeploymentPipelineSourceECR, \
     DeploymentPipelinePacoCreateThenDeployImage, DeploymentPipelineDeployECS, CodePipelineStage, CodePipelineStages, \
@@ -401,6 +402,10 @@ SUB_TYPES_CLASS_MAP = {
         'ecr_repositories': ('obj_list', ECRRepositoryPermission),
         'role_policies': ('obj_list', Policy),
         'secrets': ('str_list', PacoReference),
+        'release_phase': ('container', (DeploymentPipelineBuildReleasePhases, DeploymentPipelineBuildReleasePhase)),
+    },
+    DeploymentPipelineBuildReleasePhase: {
+        'commands': ('obj_list', DeploymentPipelineBuildReleasePhaseCommand),
     },
     DeploymentPipelineDeployCodeDeploy: {
         'minimum_healthy_hosts': ('direct_obj', CodeDeployMinimumHealthyHosts)